requirement.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
flask-migrate==4.0.4
pymysql==1.0.2
flask-bcrypt==1.0.1
python-dotenv==1.0.0
flask-cors==3.0.10